requirements.txt
requests==2.19.1
retrying==1.3.3
django==2.1rc1
celery==4.2.1
django-extensions==2.1.0
djangorestframework==3.8.2
beautifulsoup4==4.6.1
lxml==4.2.3
pytz==2018.5
regex==2018.07.11
simplejson==3.16.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
Pybooru==4.1.0
Pillow==5.2.0
moviepy==0.2.3.5
gunicorn==19.9.0
django-celery-results==1.0.1
django-celery-beat==1.1.1
django-filter==2.0.0
